03_Working/LINEBOT_RESTFUL/requirements.txt
Flask
gunicorn
line-bot-sdk
paho-mqtt
Flask
sqlalchemy
psycopg2
requests
werkzeug